pyproject.toml
[tool.poetry]
name = "zadanie-6105"
version = "0.1.0"
description = ""
authors = ["Evgenii Gusev <ged1994@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.1"
djangorestframework = "^3.15.2"
django-environ = "^0.11.2"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.9"
drf-standardized-errors = "^0.14.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"

[tool.isort]
line_length = 119
multi_line_output = 5

[tool.yapf]
align_closing_bracket_with_visual_indent = true
based_on_style = "google"
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
